refactor: changed datatypes in authentication and account classes

+++ backend/src/main/java/com/university/backend/repository/AccountRepository.java
@@ -6,9 +6,10 @@ import java.util.Optional;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccountRepository extends JpaRepository<Account, Long>
+public interface AccountRepository extends JpaRepository<Account, Integer>
 {
     Account findByEmail(String username);
     boolean existsByEmail(String email);
+
     //Optional<Account> findByEmail(String email);
 }
